requirements-gcp.txt
# GCP-specific dependencies
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
vertexai>=0.0.1

# Core application dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.3
pydantic-settings>=2.1.0
PyMuPDF>=1.24.0
pdfplumber>=0.10.4
Pillow>=10.2.0
python-json-logger>=2.0.7

# Production
gunicorn>=21.2.0
